Add new service for payment

+++ src/main/java/ar/com/estela/lavadero/repository/SalePaymentRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import ar.com.estela.lavadero.entity.SalePayment;
 
 @Repository
-public interface SalePaymentRepository extends JpaRepository<SalePayment, Integer> {
+public interface SalePaymentRepository extends JpaRepository<SalePayment, Long> {
 
 }
